+++ cli/src/voidrift_cli/phases/develop.py
@@ -100,6 +100,7 @@ def run_develop(
 
     try:
         import voidrift_mcp.server as mcp_mod
+        mcp_mod.run_id = run_id
         mcp_mod._boot()
         mcp_mod.load_tasks(str(task_file))
         tools, handlers = build_mcp_tools(mcp_mod)